# lregex: optimize substitute()