substitute was one of the bottle neck.
It called strlen before memcpy.
We can remove the strlen invocation safely.

The original code:
  $ xargs cat <  /tmp/list > /dev/null; time ./ctags -L - < /tmp/list
  ./ctags -L - < /tmp/list  97.39s user 4.72s system 97% cpu 1:45.09 total

With this change:
  $ xargs cat <  /tmp/list > /dev/null; time ./ctags -L - < /tmp/list
  ./ctags -L - < /tmp/list  87.24s user 4.68s system 97% cpu 1:34.28 total

About 10% faster!